Cast to int to avoid a gcc warning

+++ initify_plugin.c
@@ -632,7 +632,7 @@ static int get_arg_num(const gcall *call, const_tree arg)
 {
 	int idx;
 
-	for (idx = 0; idx < gimple_call_num_args(call); idx++) {
+	for (idx = 0; idx < (int)gimple_call_num_args(call); idx++) {
 		const_tree cur_arg = gimple_call_arg(call, idx);
 
 		if (cur_arg == arg)
@@ -658,7 +658,7 @@ static bool only_nocapture_call(const_tree decl)
 		int idx;
 		const gcall *call = as_a_const_gcall(e->call_stmt);
 
-		for (idx = 0; idx < gimple_call_num_args(call); idx++) {
+		for (idx = 0; idx < (int)gimple_call_num_args(call); idx++) {
 			const_tree arg = gimple_call_arg(call, idx);
 
 			if (TREE_CODE(arg) != ADDR_EXPR)
@@ -1267,7 +1267,7 @@ static void search_var_param(gcall *stmt)
 
 	pointer_set_insert(visited, stmt);
 
-	for (num = 0; num < gimple_call_num_args(stmt); num++) {
+	for (num = 0; num < (int)gimple_call_num_args(stmt); num++) {
 		const_tree type, fndecl;
 		bool has_str_cst = true;
 		tree str, arg = gimple_call_arg(stmt, num);
@@ -1301,7 +1301,7 @@ static void search_str_param(gcall *stmt)
 
 	pointer_set_insert(visited, stmt);
 
-	for (num = 0; num < gimple_call_num_args(stmt); num++) {
+	for (num = 0; num < (int)gimple_call_num_args(stmt); num++) {
 		const_tree fndecl;
 		tree str, arg = gimple_call_arg(stmt, num);
